+++ gazpacho/soup.py
@@ -2,32 +2,45 @@ from html.parser import HTMLParser
 from .utils import match, html_starttag_and_attrs
 
 class Soup(HTMLParser):
-    '''HTML Parser Class
+    '''HTML Soup Parser
 
-    - html (str): Full HTML text
-    - tag (str, None): Found tag
-    - attrs (dict, None): Found attributes
-    - text (str, None): Found text
-    - find (method): The main method to find HTML tags
-    - find_one (method): find, but for just one result
+    Attributes:
+
+    - html (str): HTML content to parse
+    - tag (str, None): HTML element tag returned by find
+    - attrs (dict, None): HTML element attributes returned by find
+    - text (str, None): HTML element text returned by find
+
+    Methods:
+
+    - find: return all matching HTML elements
+    - find_one: return the first matching HTML element
 
     Examples:
 
     ```
     from gazpacho import Soup
+
     html = "<div><p id='foo'>bar</p><p id='foo'>baz</p><p id='zoo'>bat</p></div>"
     soup = Soup(html)
-    result = soup.find('p', {'id': 'foo'})
-    print(result)
+
+    soup.find('p', {'id': 'foo'})
     # [<p id="foo">bar</p>, <p id="foo">baz</p>]
+
     result = soup.find_one('p', {'id': 'zoo'})
     print(result)
     # <p id="zoo">bat</p>
+
     print(result.text)
     # bat
     ```
     '''
+
     def __init__(self, html):
+        '''Params:
+
+        - html (str): HTML content to parse
+        '''
         super().__init__()
         self.html = html
         self.tag = None
@@ -42,7 +55,8 @@ class Soup(HTMLParser):
 
     def handle_starttag(self, tag, attrs):
         html, attrs = html_starttag_and_attrs(tag, attrs)
-        if tag == self.tag and match(self.attrs, attrs) and not self.count:
+        matching = match(self.attrs, attrs, self.strict)
+        if tag == self.tag and matching and not self.count:
             self.count += 1
             self.group += 1
             self.groups.append(Soup(''))
@@ -83,31 +97,65 @@ class Soup(HTMLParser):
         else:
             return
 
-    def find(self, tag, attrs=None):
-        '''Find all HTML elements that match a tag and optional attributes
+    def find(self, tag, attrs=None, strict=False):
+        '''Return all matching HTML elements
 
-        - tag (str): HTML tag to find
-        - attrs (dict, optional): Attributes within tag to match
+        Params:
+
+        - tag (str): HTML element tag to find
+        - attrs (dict, optional): HTML element attributes to match
+        - strict (bool, False): Require exact attribute matching
+
+        Examples:
+
+        ```
+        html = "<div><p id='foo foo-striped'>bar</p><p id='foo'>baz</p><p id='zoo'>bat</p></div>"
+        soup = Soup(html)
+
+        soup.find('p')
+        # [<p id="foo foo-striped">bar</p>, <p id="foo">baz</p>, <p id="zoo">bat</p>]
+
+        soup.find('p', {'id': 'foo'})
+        # [<p id="foo foo-striped">bar</p>, <p id="foo">baz</p>]
+
+        soup.find('p', {'id': 'foo'}, strict=True)
+        # [<p id="foo">baz</p>]
+        ```
         '''
         self.tag = tag
         self.attrs = attrs
+        self.strict = strict
         self.count = 0
         self.group = 0
         self.groups = []
-        super().feed(self.html)
-        return self.groups
+        self.feed(self.html)
+        soups = self.groups
+        return soups
 
-    def find_one(self, tag, attrs=None):
-        '''Find one HTML element that matches a tag and optional attributes
+    def find_one(self, tag, attrs=None, strict=False):
+        '''Return the first matching HTML element
 
-        - tag (str): HTML tag to find
-        - attrs (dict, optional): Attributes within tag to match
+        Params:
+
+        - tag (str): HTML element tag to find
+        - attrs (dict, optional): HTML element attributes to match
+        - strict (bool, False): Require exact attribute matching
+
+        Example (*for more see* `find`):
+
+        ```
+        html = "<div><p id='foo foo-striped'>bar</p><p id='foo'>baz</p><p id='zoo'>bat</p></div>"
+        soup = Soup(html)
+
+        soup.find_one('p', {'id': 'foo'})
+        # <p id="foo foo-striped">bar</p>
         '''
         self.tag = tag
         self.attrs = attrs
+        self.strict = strict
         self.count = 0
         self.group = 0
         self.groups = []
-        super().feed(self.html)
+        self.feed(self.html)
         soup = self.groups[0]
         return soup