+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 ### Install
 
-To install the latest version of gazpacho simply run:
+To install the latest version of gazpacho run:
 
 ```
 pip install -U gazpacho
@@ -14,5 +14,6 @@ pip install -U gazpacho
 
 #### 0.8 (2019-10-XX)
 
-* `find_one` vs. `find`
+* Split the `find` method into `find` and `find_one`
+* Enabled strict attribute matching for both `find` methods (however, defaults to `strict=False`)